requirements.txt
Flask==3.1.0
google-api-python-client==2.168.0
google-auth==2.40.1
requests==2.32.3
